refactor(objfmt): widen packReader.Len to int64
All call sites paired Len() with int64 offsets and cast on every
access. Returning int64 directly makes the contract explicit and
prevents silent truncation on 32-bit hosts holding multi-gigabyte
packs — a fallback path the type's docstring already advertises.

Wrap `golang.org/x/exp/mmap.ReaderAt` in a thin adapter that widens
its `Len() int` to int64, and drop the `int(r.size)` narrowing in
`fileReader.Len`. Call sites in `pack.go`, `pack_header.go`,
`pack_delta.go`, and `pack_verify.go` shed their `int64()` casts.

+++ internal/objfmt/mmap_reader_test.go
@@ -24,7 +24,7 @@ func TestPack_openPackReader(t *testing.T) {
 		// Satisfies the interface contract.
 		var _ io.ReaderAt = r
 		var _ io.Closer = r
-		assert.Equal(t, len(want), r.Len())
+		assert.Equal(t, int64(len(want)), r.Len())
 
 		got := make([]byte, 4)
 		n, err := r.ReadAt(got, 0)
